% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{parseLabelsFile}
\alias{parseLabelsFile}
\title{Parse Sawtooth Labels File}
\usage{
parseLabelsFile(path)
}
\arguments{
\item{path}{File path to a Sawtooth Software .labels file}
}
\value{
List of labels metadata
}
\description{
Parse Sawtooth Labels File
}
\examples{
parseLabelsFile("golf.labels")
}
